requirements.txt
altgraph==0.17
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
bleach==3.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
comtypes==1.1.9
decorator==5.0.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.7
djangorestframework==3.12.4
djongo==1.3.6
entrypoints==0.3
filelock==3.0.12
future==0.18.2
idna==2.10
ipython==7.23.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kombu==5.0.2
MarkupSafe==1.1.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
numpy==1.19.4
packaging==20.9
pandas==1.1.4
pandocfilters==1.4.3
parso==0.8.2
pefile==2019.4.18
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
Pygments==2.8.1
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pymongo==3.11.4
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.17.3
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2020.4
pywin32==300
pywin32-ctypes==0.2.0
pyzmq==22.0.3
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.2.4
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
urllib3==1.26.4
vine==5.0.0
virtualenv==20.4.3
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
webencodings==0.5.1
wikipedia==1.4.0
